🚚 rename positions folder to mng
